crypto/keybaseish/src/requirements.txt
Django==2.1.7
pycrypto==2.6.1
